+++ js/main.js
@@ -1,3 +1,2 @@
-import './util.js';
 import {getObjectOutput} from './data.js';
 getObjectOutput();
